rename ChannelTransfer
from SocketTransfer
- add generics
