--- core/src/main/java/com/local_movement/core/transfer/SocketTransfer.java
@@ -1,44 +1,0 @@
-package com.local_movement.core.transfer;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.channels.SocketChannel;
-
-public class SocketTransfer {
-
-    public static void clearRead(SocketChannel socketChannel, ByteBuffer buffer) throws IOException {
-        buffer.clear();
-        socketChannel.read(buffer);
-    }
-
-    public static void clearReadFlip(SocketChannel socketChannel, ByteBuffer buffer) throws IOException {
-        clearRead(socketChannel, buffer);
-        buffer.flip();
-    }
-
-    public static byte[] clearReadGet(SocketChannel socketChannel, ByteBuffer buffer) throws IOException {
-        clearRead(socketChannel, buffer);
-        return buffer.array();
-    }
-
-    public static void clearPut(ByteBuffer buffer, byte[] data) {
-        buffer.clear();
-        buffer.put(data);
-    }
-
-    public static void flipWrite(SocketChannel socketChannel, ByteBuffer buffer) throws IOException {
-        buffer.flip();
-        socketChannel.write(buffer);
-    }
-
-    public static void clearPutFlipWrite(byte[] data, SocketChannel socketChannel, ByteBuffer buffer) throws IOException {
-        clearPut(buffer, data);
-        flipWrite(socketChannel, buffer);
-    }
-
-    public static void writeWithMessage(byte[] message, byte[] data, SocketChannel socketChannel, ByteBuffer buffer) throws IOException {
-        SocketTransfer.clearPutFlipWrite(message, socketChannel, buffer);
-        SocketTransfer.clearPutFlipWrite(data, socketChannel, buffer);
-    }
-
-}
